moved right rail to shared component
