+++ src/pages/research/online-resources/[slug].js
@@ -17,7 +17,7 @@ import { withRedux } from './../../../redux/withRedux';
 // Components
 import { SkeletonLoader } from '@nypl/design-system-react-components';
 import PageContainer from './../../../components/shared/layouts/PageContainer';
-import RightRail from './../../../components/location-finder/RightRail';
+import RightRail from './../../../components/shared/RightRail';
 import SearchForm from './../../../components/online-resources/SearchForm';
 import OnlineResourceCard from './../../../components/online-resources/OnlineResourceCard';
 import SearchHeader from './../../../components/shared/SearchHeader';
