+++ src/styles/main.scss
@@ -25,6 +25,5 @@
 @import './../components/location-finder/Location/Location.scss';
 @import './../components/location-finder/Map/Map.scss';
 @import './../components/location-finder/BottomPromo/BottomPromo.scss';
-@import './../components/location-finder/RightRail/RightRail.scss';
 // DS Overrides
 @import './ds-overrides.scss';
